Removed define/public and define/private from class/0.
Since last year we've been meaning to do this.  So define/public and
define/private have been removed.  This will break some old class/0
programs, but since the class has moved on to subsequent languages,
this shouldn't be too much of a pain.

The book still contains define/public and define/private uses that
need to be removed.